Providing more hooks (onUnpinned, onPinned, onUnfixed) into the lifecycle of headroom component. These callbacks are called after the state has changed. Previous hooks (onUnpin, onPin, onUnfix) are actually called before the state change. Added new parameter ‘forceState’ for forcing the state of the component explicitly from calling component. Combined with these changes we can achieve a use case where we want multiple headrooms to be displayed with one at a time only. Currently without any change in the core library we can achieve multiple headrooms but they all will be rendered on top of each other which is fine if we don’t have any box shadow effect on them. With box shadow effect the shadow becomes darker because of all pinned headrooms, which might not look good to end user.
Caution: While attaching the callbacks(new or old) with the component, one has to keep in mind that callback should not change ‘forceState’ param such that it creates infinite event loop e.g. setting forceState to be pinned in onPinned or onPin handler. We might improve the library in future to overcome(may be by avoiding calls to these callbacks when state is externally forced) this situation instead of delegating it to the caller.